Mods from AP Hill

+++ client_libs/libplayerc/dev_truth.c
@@ -101,8 +101,7 @@ void playerc_truth_putdata(playerc_truth_t *device, player_msghdr_t *header,
 
   device->px = ntohl(data->px) / 1000.0;
   device->py = ntohl(data->py) / 1000.0;
-  //device->pa = ntohl(data->pa) * M_PI / 180;
-  device->pa = ((double)(int32_t)ntohl(data->pa)) * M_PI / 180;
+  device->pa = ((double) (int32_t) ntohl(data->pa)) * M_PI / 180;
 }
 
 
@@ -124,10 +123,9 @@ int playerc_truth_get_pose(playerc_truth_t *device, double *px, double *py, doub
     return -1;
   }
 
-  *px = ntohl(config.px) / 1000.0;
-  *py = ntohl(config.py) / 1000.0;
-  //*pa = ntohl(config.pa) * M_PI / 180;
-  *pa = ((double)(int32_t)ntohl(config.pa)) * M_PI / 180;
+  *px = ((int32_t) ntohl(config.px)) / 1000.0;
+  *py = ((int32_t) ntohl(config.py)) / 1000.0;
+  *pa = ((int32_t) ntohl(config.pa)) * M_PI / 180;
 
   return 0;
 }
@@ -140,9 +138,9 @@ int playerc_truth_set_pose(playerc_truth_t *device, double px, double py, double
   player_truth_pose_t config;
 
   config.subtype = PLAYER_TRUTH_SET_POSE;
-  config.px = htonl((int) (px * 1000));
-  config.py = htonl((int) (py * 1000));
-  config.pa = htonl((int) (pa * 180 / M_PI)); 
+  config.px = htonl((uint32_t) (int32_t) (px * 1000));
+  config.py = htonl((uint32_t) (int32_t) (py * 1000));
+  config.pa = htonl((uint32_t) (int32_t) (pa * 180 / M_PI)); 
   
   len = playerc_client_request(device->info.client, &device->info,
                                &config, sizeof(config), &config, sizeof(config));
